automatable deploy

+++ deploy.sh
@@ -1,8 +1,16 @@
-printf "%s" "release tag: "
-read tag
+set -e
+
+tag=`gh release list -R hatchdotlol/hatch-api --json name | jq -r ".[0].name"`
+if [ $(ls /home/server/api | grep "$tag") != "" ]; then
+        exit 0
+fi
+
 rm -rf api/*
 cd api
-gh release download "$tag" --clobber -R hatchdotlol/api
+gh release download "$tag" -p "hatch-api_*_x86_64-unknown-linux-musl.zip" --clobber -R hatchdotlol/api
 unzip *_x86_64-unknown-linux-musl.zip
-./api &
+mv ./api "./api_$tag"
+rm *.zip
+pkill api*
+./api* &
 cd ..
